@@ -51,7 +51,6 @@ setup(
         'Programming Language :: Python',
         'Programming Language :: Python :: 3 :: Only',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
@@ -70,9 +69,9 @@ setup(
         'pytyped': ['py.typed'],
     },
     install_requires=[
-        'msrest>=0.6.21',
+        'msrest>=0.7.1',
         'azure-common~=1.1',
-        'azure-mgmt-core>=1.3.1,<2.0.0',
+        'azure-mgmt-core>=1.3.2,<2.0.0',
     ],
-    python_requires=">=3.6"
+    python_requires=">=3.7"
 )
