@@ -9,12 +9,11 @@
 from copy import deepcopy
 from typing import Any, Awaitable, TYPE_CHECKING
 
-from msrest import Deserializer, Serializer
-
 from azure.core.rest import AsyncHttpResponse, HttpRequest
 from azure.mgmt.core import AsyncARMPipelineClient
 
 from .. import models
+from .._serialization import Deserializer, Serializer
 from ._configuration import CommunicationServiceManagementClientConfiguration
 from .operations import CommunicationServicesOperations, DomainsOperations, EmailServicesOperations, Operations
 
@@ -22,7 +21,8 @@ if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from azure.core.credentials_async import AsyncTokenCredential
 
-class CommunicationServiceManagementClient:
+
+class CommunicationServiceManagementClient:  # pylint: disable=client-accepts-api-version-keyword
     """REST API for Azure Communication Services.
 
     :ivar operations: Operations operations
@@ -34,13 +34,13 @@ class CommunicationServiceManagementClient:
     :vartype domains: azure.mgmt.communication.aio.operations.DomainsOperations
     :ivar email_services: EmailServicesOperations operations
     :vartype email_services: azure.mgmt.communication.aio.operations.EmailServicesOperations
-    :param credential: Credential needed for the client to connect to Azure.
+    :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials_async.AsyncTokenCredential
-    :param subscription_id: The ID of the target subscription.
+    :param subscription_id: The ID of the target subscription. Required.
     :type subscription_id: str
     :param base_url: Service URL. Default value is "https://management.azure.com".
     :type base_url: str
-    :keyword api_version: Api Version. Default value is "2021-10-01-preview". Note that overriding
+    :keyword api_version: Api Version. Default value is "2022-07-01-preview". Note that overriding
      this default value may result in unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
@@ -54,7 +54,9 @@ class CommunicationServiceManagementClient:
         base_url: str = "https://management.azure.com",
         **kwargs: Any
     ) -> None:
-        self._config = CommunicationServiceManagementClientConfiguration(credential=credential, subscription_id=subscription_id, **kwargs)
+        self._config = CommunicationServiceManagementClientConfiguration(
+            credential=credential, subscription_id=subscription_id, **kwargs
+        )
         self._client = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
@@ -62,16 +64,13 @@ class CommunicationServiceManagementClient:
         self._deserialize = Deserializer(client_models)
         self._serialize.client_side_validation = False
         self.operations = Operations(self._client, self._config, self._serialize, self._deserialize)
-        self.communication_services = CommunicationServicesOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.communication_services = CommunicationServicesOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
         self.domains = DomainsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.email_services = EmailServicesOperations(self._client, self._config, self._serialize, self._deserialize)
 
-
-    def _send_request(
-        self,
-        request: HttpRequest,
-        **kwargs: Any
-    ) -> Awaitable[AsyncHttpResponse]:
+    def _send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
@@ -80,7 +79,7 @@ class CommunicationServiceManagementClient:
         >>> response = await client._send_request(request)
         <AsyncHttpResponse: 200 OK>
 
-        For more information on this code flow, see https://aka.ms/azsdk/python/protocol/quickstart
+        For more information on this code flow, see https://aka.ms/azsdk/dpcodegen/python/send_request
 
         :param request: The network request you want to make. Required.
         :type request: ~azure.core.rest.HttpRequest
